requirements.txt
alembic==1.13.1
async-timeout==4.0.3
bidict==0.22.1
blinker==1.7.0
cachelib==0.9.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-engineio==4.8.2
python-socketio==5.11.0
redis==5.0.1
simple-websocket==1.0.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
wsproto==1.2.0
